refactor(mcp-adapters): minimize SDK migration changes
