+++ libs/langchain-mcp-adapters/src/client.ts
@@ -9,7 +9,6 @@ import type {
 import { StdioClientTransport } from "@modelcontextprotocol/client/stdio";
 import type { DynamicStructuredTool } from "@langchain/core/tools";
 import { z } from "zod/v3";
-import { z as z4 } from "zod/v4";
 import { loadMcpTools } from "./tools.js";
 import { ConnectionManager, type Client } from "./connection.js";
 import { getDebugLog } from "./logging.js";
@@ -30,13 +29,6 @@ import {
   type LoadMcpToolsOptions,
   _resolveAndApplyOverrideHandlingOverrides,
 } from "./types.js";
-
-const httpStatusSchema = z4.number().int().min(100).max(599);
-const httpErrorSchema = z4.object({
-  status: httpStatusSchema.optional().catch(undefined),
-  code: httpStatusSchema.optional().catch(undefined),
-  message: z4.string().optional().catch(undefined),
-});
 
 const debugLog = getDebugLog();
 
@@ -763,18 +755,24 @@ export class MultiServerMCPClient {
   }
 
   private _getHttpErrorCode(error: unknown): number | undefined {
-    const parsed = httpErrorSchema.safeParse(error);
-    if (!parsed.success) return undefined;
+    if (typeof error !== "object" || error === null) return undefined;
+
+    const isHttpStatus = (value: unknown): value is number =>
+      typeof value === "number" &&
+      Number.isInteger(value) &&
+      value >= 100 &&
+      value <= 599;
 
     // SDK 2 HTTP errors use status; SSE errors use a numeric code.
-    const { status, code, message } = parsed.data;
-    if (status !== undefined) return status;
-    if (code !== undefined) return code;
+    if ("status" in error && isHttpStatus(error.status)) return error.status;
+    if ("code" in error && isHttpStatus(error.code)) return error.code;
 
-    const match = message?.match(/\(HTTP (\d{3})\)/);
-    if (!match) return undefined;
-    const parsedStatus = httpStatusSchema.safeParse(Number(match[1]));
-    return parsedStatus.success ? parsedStatus.data : undefined;
+    if (!("message" in error) || typeof error.message !== "string") {
+      return undefined;
+    }
+    const match = error.message.match(/\(HTTP (\d{3})\)/);
+    const status = match ? Number(match[1]) : undefined;
+    return isHttpStatus(status) ? status : undefined;
   }
 
   private _createAuthenticationErrorMessage(
